Fix `<figure>` elements being placed inside `<p>` elements
Paragraph elements cannot contain figure elements as figures are not
considered "Phrasing content".

This fix ensures that in simple cases where an image is in a paragraph
and that paragraph is inside a non-paragraph element, the figure will be
placed after the paragraph, not within it.

Doing so prevents the plugin from outputting invalid HTML.